Open-sourcing the gamelib library I wrote/use.
This is rough code, and contains lots of cruft. I'll work on making
it better, but be warned this was opened-source without reviewing
for bad practices.

Side note, the gamelib-net-* stuff is in work. I was bored and
wanted to rewrite a generic client/server communication library as
a way to learn netty.